Initial commit for SudokuSolver.
TestRun.groovy

Simple temporary entry point for the script. Currently contains 2 test cases for Sudoku puzzles 1 of easy difficulty and 1 of medium difficulty.

SudokuSlot.groovy

Represents a single square on a Sudoku puzzle. Stores possible values that can belong in this slot.

SudokuSlotCollection.groovy

As the name suggests, represents a collection of slots. Purpose is to be used to represent columns, rows, and individual "blocks" of the sudoku puzzle. SudokuSlotCollection stores if the collection has been solved or not, this helps with performance to avoid recalculating finished segments of the puzzle.

SudokuBoard.groovy

Represents the entire Sudoku puzzle. contains the majority of the logic for solving the puzzle, as well as various console printout methods for initial debugging purposes