Fixed an incorrect value in test_puzzle1(easy).txt Added a new Medium difficulty test file Added a hard difficulty test file
Added a new SudokuSlot constructor to take in a value for the slot on creation.

Added a calculatePossibleValueOccurrences method to the SudokuSlotCollection class which calculates how many occurrences  of each number is in the possibleValue lists of all of the slots in the collection.

Updated the logic in the SudokuBoard.groovy class to increase accuracy and completeness of Sudoku solutions. Logic is not yet complete, but seems to work better for medium difficulty puzzles