Added a new puzzle file (test_puzzle6(medium)
Created 2 new classes, SudokuSlotArray, and SudokuBlock. SudokuSlotArray represents columns and rows in a sudoku puzzle. SudokuBlock represents a block in the sudoku puzzle. Reworked the logic so most of the work  occurs in the SudokuSlotArray and SudokuBlock classes.

The script now handles calculating which values can possibly be assigned to every slot, how many times each possible value occurs in each row, column, and block, and will assign a value to a slot if it is the only value that can belong in that slot, or if it is the only slot that value can belong to. This seems to be able to solve most (if not all) easy puzzles, and solves some medium puzzles.